+++ setup.cfg
@@ -19,10 +19,10 @@ per-file-ignores =
 strict_optional = True
 ignore_missing_imports = True
 files =
-    flytekit/annotated/**/*.py,
+    flytekit/core/**/*.py,
     plugins/**/*.py,
     flytekit/types/**/*.py,
-    tests/flytekit/unit/annotated/**/*.py
+    tests/flytekit/unit/core/**/*.py
 follow_imports = skip
 cache_dir = /dev/null
